Move Header and Footer into own directories
